refactor: merged git.service into git.provider
