+++ apps/api/src/git/git.module.ts
@@ -1,9 +1,8 @@
 import { Module } from '@nestjs/common';
 import { GithubService } from './github.service';
-import { GitService } from './git.service';
 
 @Module({
-  providers: [GitService, GithubService],
-  exports: [GitService, GithubService],
+  providers: [GithubService],
+  exports: [GithubService],
 })
 export class GitModule {}
